[Utils][Bugfix] Provide full path to shutil.copy in copy_tokenizer
Prior to this commit, `copy_tokenizer` provided the
`$ARTIFACT_DIR/params` directory as the destination argument to
`shutil.copy`.  If the directory already exists, then this would work
correctly.  If the directory did not already exist, then it would
create a file named `params` within the artifact directory.  This code
path could be triggered by running `build.py` first with
`--build-model-only`, and later with `--convert-weights-only`.

This commit updates `copy_tokenizer` to create the `params` directory
if it does not already exist, and to provide a full output path to
`shutil.copy` instead of an output directory.

In addition, the `utils.save_params` method has an explicit check to
ensure that any previously existing `$ARTIFACT_DIR/params` is a
directory, to throw an error earlier if not.  (This was the location
of the first observed error when running with `--convert-weights-only`
after running `--build-model-only`.)