[FP8] Perform calibration in fp16
This updates the calibration flow. For calibration, weights are
quantized to fp8 at compile time. During calibration runs, both weights
and activations are dequantized to fp16 to compute gemm / group gemm and
then compute the scale. Note that activation are not explicitly
dequantized, instead we rely on alpha-scaling of fp16 group gemm.