Update worker.py for compatibility with upstream TVM
This commit updates `mlc_llm.cli.worker` to be compatible with
upstream TVM https://github.com/apache/tvm/pull/17180, which adds a
`num_groups` argument to the disco worker function.

To de-couple this compatibility from a general TVM version bump, this
commit has a check on the number of `worker.py` arguments provided, to
determine whether the `num_groups` argument is present.  After the TVM
version used by MLC-LLM is updated to include the upstream changes,
this check can be removed.